requirements.txt
# Core dependencies
streamlit>=1.28.0
groq>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.5.0

# Weather API
openmeteo-requests>=1.1.0
requests-cache>=1.1.0
retry-requests>=2.0.0

# Data processing
numpy>=1.26.0
pandas>=2.1.0
